ResizeReversePaginatedWebView fails on some OS versions on some hardware https://bugs.webkit.org/show_bug.cgi?id=224699 <rdar://76782199>
Reviewed by Megan Gardner.

* TestWebKitAPI/Tests/WebKit/ResizeReversePaginatedWebView.cpp:
(TestWebKitAPI::didLayout):
This test does *not* care about the specific number of pages that are
generated; the key to the test is the scroll position check.

Since the page-count check in this test has caused lots of trouble
in the past (see the history) just remove it and replace it with a
check that we ended up paginated at all.


Canonical link: https://commits.webkit.org/237677@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@277429 268f45cc-cd09-0410-ab3c-d52691b4dbfc