[iOS] REGRESSION(r271216) pointerevents/ios/pointer-events-no-mousedown-when-prevent-default-called-on-pointerdown.html is consistently failing https://bugs.webkit.org/show_bug.cgi?id=225734 <rdar://problem/77070765>
Reviewed by Tim Horton.

r271216 made it so that `touchWithIdentifierWasRemoved` is eagerly called when dispatching
`"pointerup"` for a touch event. This is problematic because (compatibility) mouse events
are dispatched after both pointer and touch events, meaning that if the `CapturingData` for
the touch event is removed, there's no way of knowing if the pointer event `preventDefault`.

Instead of fully removing the `CapturingData` entirely, change its state such that it's
marked as `CapturingData::State::Finished` and check for that instead.

* page/PointerCaptureController.h:
* page/PointerCaptureController.cpp:
(WebCore::PointerCaptureController::hasCancelledPointerEventForIdentifier const):
(WebCore::PointerCaptureController::dispatchEventForTouchAtIndex):
(WebCore::PointerCaptureController::pointerEventForMouseEvent):
(WebCore::PointerCaptureController::cancelPointer):


Canonical link: https://commits.webkit.org/237706@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@277463 268f45cc-cd09-0410-ab3c-d52691b4dbfc