[MediaStream][GStreamer] Flaky fast/mediastream/MediaStream-video-element-video-tracks-disabled.html https://bugs.webkit.org/show_bug.cgi?id=225651
Reviewed by Alicia Boya Garcia.

Source/WebCore:

Push the black frame as soon as the corresponding track has been disabled. The black frame
is also now created once only and reused, instead of re-creating it 30 times per second (or
whatever the frame rate is). The cached frame is updated when the track dimensions change as
well.

* platform/mediastream/gstreamer/GStreamerMediaStreamSource.cpp:

LayoutTests:

* fast/mediastream/MediaStream-video-element-video-tracks-disabled.html: Delay the
notifyDone call to give the video element some time to render the black frame corresponding
to the disabled track. This might still flake on the bots, if that is the case we might need
to reassess how the test behaves or decide to mark it as expected flaky on glib ports.

Canonical link: https://commits.webkit.org/237857@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@277651 268f45cc-cd09-0410-ab3c-d52691b4dbfc