[css-flexbox] Wrong height of an empty table inside an orthogonal flex parent https://bugs.webkit.org/show_bug.cgi?id=225339
Patch by Felipe Erias <felipeerias@igalia.com> on 2021-05-20
Reviewed by Sergio Villar Senin.
LayoutTests/imported/w3c:

Add a test for an empty table inside a flexbox container, where the logical heights of both elements
are orthogonal to each other (the table has vertical writing mode and the flexbox uses row direction).

This is a flipped version of table-as-item-specified-height.html.

* web-platform-tests/css/css-flexbox/table-as-item-specified-width-vertical-expected.html: Added.
* web-platform-tests/css/css-flexbox/table-as-item-specified-width-vertical.html: Added.

Source/WebCore:

When setting the height of an empty table, give precedence to the overriding value if it has been set.
Do not cache the height of that empty table as the intrinsic content height, because doing so may cause
layout problems (the table is actually empty and may be assigned a different height by its parent).

Test: imported/w3c/web-platform-tests/css/css-flexbox/table-as-item-specified-height.html
      imported/w3c/web-platform-tests/css/css-flexbox/table-as-item-specified-width-vertical.html

* rendering/RenderTable.cpp:
(WebCore::RenderTable::layout):

LayoutTests:

* TestExpectations: Remove one WPT test that now passes.

Canonical link: https://commits.webkit.org/237938@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@277777 268f45cc-cd09-0410-ab3c-d52691b4dbfc