The test of SmallSet.cpp should pass the right HashTrait to the HashSet it uses as a reference https://bugs.webkit.org/show_bug.cgi?id=226027
Reviewed by Mark Lam.

SmallSet allows all values except for the max value.
So we test it on small values, including 0.

HashSet by default forbids 0, so it made the test crash.
We can trivially fix it by passing UnsignedWithZeroKeyHashTraits, which allows everything which is not max() or max()-1

* TestWebKitAPI/Tests/WTF/SmallSet.cpp:
(TestWebKitAPI::testSmallSetOfUnsigned):


Canonical link: https://commits.webkit.org/237964@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@277817 268f45cc-cd09-0410-ab3c-d52691b4dbfc