[ wk2 ] http/tests/security/contentSecurityPolicy/report-only-connect-src-xmlhttprequest-redirect-to-blocked.py is a constant text failure https://bugs.webkit.org/show_bug.cgi?id=224626 <rdar://problem/76729427>
http/tests/security/contentSecurityPolicy/report-only-connect-src-beacon-redirect-blocked.py
initiates a redirect that generates a CSP violation, but does not wait
for the console logging. This results in flakiness of other test that
occasionally pick up this console logging in their output. To fix this,
we should wait to end the test until we get the CSP violation report.

Reviewed by Sam Weinig.

* http/tests/security/contentSecurityPolicy/report-only-connect-src-beacon-redirect-blocked-expected.txt:
Update expectations now that we consistently expect to get report
violation console logging for this test.

* http/tests/security/contentSecurityPolicy/report-only-connect-src-beacon-redirect-blocked.py:
* platform/wk2/TestExpectations:


Canonical link: https://commits.webkit.org/238079@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@277952 268f45cc-cd09-0410-ab3c-d52691b4dbfc