REGRESSION: ASSERTION FAILED: !isCompilationThread() in JSC::Profiler::Database::addCompilation() https://bugs.webkit.org/show_bug.cgi?id=226302 <rdar://78537378>
Reviewed by Saam Barati.

We started linking the JIT code in the compiler thread, which now triggers this assertion.
The assertion was introduce before a lock was held to modify the data structures in the
profiler database, but it's safe to remove it now.

* profiler/ProfilerDatabase.cpp:
(JSC::Profiler::Database::addCompilation):


Canonical link: https://commits.webkit.org/238184@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278139 268f45cc-cd09-0410-ab3c-d52691b4dbfc