Remove selective compilation of libwebrtc as TestWebKitAPI depends on it for boringssl, so it is always needed https://bugs.webkit.org/show_bug.cgi?id=226430
Reviewed by Tim Horton.

r244568 made TestWebKitAPI require linking the built boringssl from
the libwebrtc, so it really can't be selectively compiled. Over time,
we have re-enabled essentially all the platforms, when the intent of
disabling it at all originally was to avoid double building it for
Mac Catalyst, which we retain.

* Makefile:


Canonical link: https://commits.webkit.org/238284@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278247 268f45cc-cd09-0410-ab3c-d52691b4dbfc