Improve if condition in RenderReplaced::computeReplacedLogicalWidth https://bugs.webkit.org/show_bug.cgi?id=226325
Reviewed by Darin Adler.

Removed the check for intrinsicRatio as we're already inside a block in which intrinsicRatio
is true. Also switched the position of the two conditions in the OR clause as it allows us
to remove the check for computedHeightIsAuto. Last but not least, used a ternary operator
to initialize the value of boxSizing instead of the if clause.

No new tests as there is no change in functionality.

* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::computeReplacedLogicalWidth const):

Canonical link: https://commits.webkit.org/238337@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278300 268f45cc-cd09-0410-ab3c-d52691b4dbfc