[Cocoa] Clean up notification_center message filters https://bugs.webkit.org/show_bug.cgi?id=226555 <rdar://problem/77164076>
Reviewed by Per Arne Vollan.

We initially had two groups of mach-message-send filter rules for 'com.apple.system.notification_center' so we could
collect telemetry on a group of messages. Now that we have removed all the messages we do not need, there is no reason
to maintain two separate sets.

This patch just combines the two sets into a single list so we can easily conditionalize it for different OS releases
if needed.

* Resources/SandboxProfiles/ios/com.apple.WebKit.WebContent.sb.in:
* WebProcess/com.apple.WebProcess.sb.in:


Canonical link: https://commits.webkit.org/238410@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278383 268f45cc-cd09-0410-ab3c-d52691b4dbfc