Flaky assertion hit in AudioSessionRoutingArbitratorProxy::endRoutingArbitration() https://bugs.webkit.org/show_bug.cgi?id=226699
Reviewed by Darin Adler.

If m_setupArbitrationOngoing is true, then beginRoutingArbitrationForToken() only adds the token to
m_tokens once the setup is done, asynchronously. As a result, if endRoutingArbitrationForToken() during
setup, the `isInRoutingArbitrationForToken(token)` assertion will hit, because the token wasn't added
to m_tokens yet. Tweak the assertion so that it doesn't hit when m_setupArbitrationOngoing is true.

This is a speculative fix as I wasn't able to reproduce the assertion hit locally.

* platform/audio/mac/SharedRoutingArbitrator.mm:
(WebCore::SharedRoutingArbitrator::endRoutingArbitrationForToken):


Canonical link: https://commits.webkit.org/238539@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278541 268f45cc-cd09-0410-ab3c-d52691b4dbfc