[WTF][GStreamer] Fix clang TSA warnings in WTF::DataMutex https://bugs.webkit.org/show_bug.cgi?id=226719
Reviewed by Xabier Rodriguez-Calvar.

Fix the remaning clang thread safety warnings in WTF::DataMutex.

The goal of this patch is to reduce the number of warnings in the
GStreamer codebase. Whether DataMutex should be deprecated in favor of
Locker with the clang TSA annotations is outside of the scope of this
patch.

* wtf/DataMutex.h:


Canonical link: https://commits.webkit.org/238556@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278558 268f45cc-cd09-0410-ab3c-d52691b4dbfc